Add global design token docs
