+++ aries-site/src/components/content/DesignTokenContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const DesignTokenContext = createContext({});
