+++ aries-site/src/components/content/index.js
@@ -4,6 +4,8 @@ export * from './ColorCompliance';
 export * from './ColorSwatch';
 export * from './ComingSoon';
 export * from './DecisionTree';
+export * from './DesignTokensTable';
+export * from './DesignTokenContext';
 export * from './DocsMessage';
 export * from './HighlightPhrase';
 export * from './PageBackground';
@@ -12,3 +14,4 @@ export * from './Status';
 export * from './SubmitFeedback';
 export * from './ThemeModeToggle';
 export * from './MarkdownComponents';
+export * from './designTokenUtils';
