+++ aries-site/src/layouts/content/ContentPane.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Box } from 'grommet';
+
+export const ContentPane = ({ ...rest }) => (
+  <Box background="background-front" pad="medium" round="medium" {...rest} />
+);
